Refactor - change directory name from symphony to main, due to conflict during import
